Handle unsigned devicectl CPU subtypes
Read Xcode 27 CPU subtype values as UInt64 so sign-extended ARM64e values retain their low-byte architecture discriminator. Update the v5 regression fixture to use the exact unsigned value emitted by devicectl.


Copilot-Session: 891a4a77-04cb-4497-8c29-83759604e389

+++ Xamarin.MacDev/DeviceCtlOutputParser.cs
@@ -129,7 +129,7 @@ public static class DeviceCtlOutputParser {
 			return "x86_64";
 		case 0x0100000c:
 			if (cpuType.Value.TryGetProperty ("subtype", out var subtypeElement) &&
-				subtypeElement.TryGetInt32 (out var subtype) &&
+				subtypeElement.TryGetUInt64 (out var subtype) &&
 				(subtype & 0xff) == 2)
 				return "arm64e";
 			return "arm64";
